pyproject.toml
[tool.poetry]
name = "quack2tex"
version = "0.3.0"
description = "Converts screenshots of text to LaTeX, and more using LLMs."
authors = ["haruiz <henryruiz22@gmail.com>"]
readme = "README.md"
packages = [{include = "quack2tex", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
mss = "^9.0.1"
pillow = "^10.4.0"
matplotlib = "^3.9.1.post1"
google-generativeai = "^0.7.2"
python-dotenv = "^1.0.1"
pyside6 = "^6.7.2"
markdown = "^3.6"
pydantic = "^2.8.2"
openai = "^1.40.6"
ollama = "^0.3.1"
tenacity = "^9.0.0"
typer = "^0.12.3"
scipy = "^1.14.1"
pydantic-settings = "^2.5.2"
sqlalchemy = "^2.0.34"
aiosqlite = "^0.20.0"

[tool.poetry.scripts]
quack2tex = "quack2tex.cli:run"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
